备份设置脚本

+++ setup-router/acme.sh/acme-remote-deploy.sh
@@ -2,13 +2,7 @@
 
 DOMAIN_NAME=owent.net
 ACMESH_SSL_DIR=/data/website/ssl
-if [[ -e "$ACMESH_SSL_DIR/${DOMAIN_NAME}_ecc" ]]; then
-  INSTALL_CERT_DIR="$ACMESH_SSL_DIR/${DOMAIN_NAME}_ecc"
-else
-  INSTALL_CERT_DIR="$ACMESH_SSL_DIR/${DOMAIN_NAME}"
-fi
 REMOTE_DEPLOY_KEY=<path of id_ed25519>
-REMOTE_DEPLOY_SSL_PATH=$INSTALL_CERT_DIR
 
 # Update local services
 if [[ "x$1" == "xupdate-v2ray" ]] || [[ "x$1" == "xupdate-vproxy" ]] || [[ "x$1" == "xupdate-vbox" ]]; then
@@ -30,12 +24,19 @@ for NODE_ADDR in ${REPLICATION_NODES[@]}; do
 
     echo "============ Upload SSL files to $NODE_USER@$NODE_HOST:$NODE_PORT ... " 
 
-    ssh -p $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i $REMOTE_DEPLOY_KEY "$NODE_USER@$NODE_HOST" "mkdir -p $REMOTE_DEPLOY_SSL_PATH" 
-    scp -P $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i \
-      "$REMOTE_DEPLOY_KEY" "$INSTALL_CERT_DIR/$DOMAIN_NAME.cer"                                         \
-      "$INSTALL_CERT_DIR/$DOMAIN_NAME.key" "$INSTALL_CERT_DIR/fullchain.cer"                            \
-      "$NODE_USER@$NODE_HOST:$REMOTE_DEPLOY_SSL_PATH"
-    ssh -p $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i "$REMOTE_DEPLOY_KEY" "$NODE_USER@$NODE_HOST" "cd $REMOTE_DEPLOY_SSL_PATH && chmod 640 *"
+    for INSTALL_CERT_DIR in "$ACMESH_SSL_DIR/${DOMAIN_NAME}_ecc" "$ACMESH_SSL_DIR/${DOMAIN_NAME}"; do
+        if [[ ! -e "$INSTALL_CERT_DIR/$DOMAIN_NAME.cer" ]]; then
+            continue
+        fi
+        REMOTE_DEPLOY_SSL_PATH=$INSTALL_CERT_DIR
+        ssh -p $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i $REMOTE_DEPLOY_KEY "$NODE_USER@$NODE_HOST" "mkdir -p $REMOTE_DEPLOY_SSL_PATH" 
+        scp -P $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i \
+          "$REMOTE_DEPLOY_KEY" "$INSTALL_CERT_DIR/$DOMAIN_NAME.cer"                                         \
+          "$INSTALL_CERT_DIR/$DOMAIN_NAME.key" "$INSTALL_CERT_DIR/fullchain.cer"                            \
+          "$NODE_USER@$NODE_HOST:$REMOTE_DEPLOY_SSL_PATH"
+        ssh -p $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i "$REMOTE_DEPLOY_KEY" "$NODE_USER@$NODE_HOST" "cd $REMOTE_DEPLOY_SSL_PATH && chmod 640 *"
+    done
+    # ssh -p $NODE_PORT -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o User=$NODE_USER -i "$REMOTE_DEPLOY_KEY" "$NODE_USER@$NODE_HOST" "cd $REMOTE_DEPLOY_SSL_PATH && chmod 640 *"
 done
 
 # Update remote services
